requirements.txt
selenium
pytest
pytest-html
pytest-xdist
openpyxl
webdriver-manager
allure-pytest
